Add List All Model Versions Query (#73)
* Add ListAllModelVersions Query

* Add list all model versions query on frontend

* Encrypt storage provider secrets (#72)

Resolves #26

I hate cryptography and existence is pain. This adds encryption
to storage provider secrets using ChaCha20-Poly1305. Decryption
of the Access Key ID is automatic when returning the value to
a GQL client, and the secret access key is never displayed to
the client again after being entered, although will be accessed
and decrypted as needed to perform operations against the upstream
storage provider.

There's an outstanding issue with adding the secret key as an
environment variable in the k8s manifest that I haven't resolved
yet, so I'm punting on it for the moment and have a temporary and

    IMPORTANT: TOTALLY INSECURE (intentionally)

key hardcoded in the encryption utility itself. Obviously this is
A Bad Thing (tm) and is just a stopgap until I get my act together.

* ZSH cross-shell compatibility for storage script (#74)

* Port Nexus to Pothos (#76)

* Nexus->Pothos port

* working storage provider methods

* ported model methods

* model version ports

* presigned URL port

* clean up an eslint comment

* Add Current Model Version Patch

* Add updated current model version query to frontend

* Fix MLModel object type

* Remove UUID extension in currentModelVersion patch

---------

Co-authored-by: Christopher Wetherill <1910269+wetherc@users.noreply.github.com>